+++ packages/iti-react/src/hooks/useCtrlEventListener.ts
@@ -3,6 +3,7 @@ import useEventListener from '@use-it/event-listener'
 // Use this to allow the user to submit a form by pressing Ctrl+Enter
 // or Command+Enter on Mac
 export function useCtrlEnterListener(callback: () => unknown, enabled: boolean): void {
+    // keypress will not work
     useEventListener<'keydown'>('keydown', (e) => {
         if (!enabled) return